Move listener tests into listener_test.go
